Rename old buy actions
